safari-600.7-branch/LayoutTests/java/lc3/ConvertJSObject/ToShort-002-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  JavaScript Object to int

FAIL TEST_CLASS.ambiguous( MyFunction ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( MyOtherFunction ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new MyObject(12345) ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new MyOtherObject("12345") ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new AnotherObject("12345") ) +'' should be SHORT. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
